shortcuts: import shortcut engine
This adds my gtk-shortcut-engine into Builder using the Ide namespace.
There is a copy_and_paste.py script in the shortcut-engine prototype
repository to import the sources into a project (in this case Builder)
with a given prefix. I find this cleaner, for now, than creating yet
another library that plugins and things would need to link against.

As an asside, I'd like to make a lot of things just be in Builder and
get down to a single binary, but that will take a lot of work.

Anyway, this does not get things wired up into Builder, it only imports
the project and gets things compiling and introspection working.