buffer: add IdeBuffer::change-settled
This signal is a convenience signal to do background work when
the user has stopped interactively typing. We will probably
tweak this as we go on but what is here now mimics what the
diagnosis engine was doing.

You should use this if you want to update background work based
on changes to the editor.