project-tree: style things similar to the mockup
While we will still likely use the DzlStackList for the default
implementation of the project-tree, or GtkTreeView based version can still
be rather sleek if we tweak some CSS and properties.

This removes the expander column since all of our expansions look like
folders anyway. We tweak the spacing and use level-indentation to get
the same indentation effect per-level.