layout: add support for specifying foreground color
This allows us to get a better set of colors for the layout
stack based on the document content.

There are a few tricks in here. We generate a unique CSS
identifier for the layout stack so that we can select on it.
It is just a monotonic number that is incremented for each
instance of layout stack that is loaded.

Because we need to inform our children widgets, the CSS is
connected globally, not to the layoutstack itself.