editor: use properties group from dazzle
This still doesn't work for our "insert-spaces-instead-of-tabs" case,
but in all honestly, I consider that a Gtk but in that it's action
helper does not know how to deal with stateful vs actionable GAction.

We need to work around it like we did previously (which was to create
our own stateful GSimpleAction and keep it up to date).