editor: start on new document button
Now that all the features are in Dazzle for this, we can simplify things
here. We still have a bunch of style tweaks to do, for example, the
design calls for the button to be closer to the perpsectives button.

Also, the styling of the menu is still a bit off.