menus: use DzlMenuButton for run menu button
This is yet another menu we can consolidate. It also allows for
plugins to specify their extensions to the menu without strictly
implementing a run handler (event if it does currently require
duplicating the icon/name entries).

What might be a possibility going forward is to generate the
GMenu entries based on the handler registration.