layout-view: add API to report errors
This at least gives us a way to report errors to the user even
if it isn't how we want to do things long term.

This is still one downside to this, if set_failed() is set to
TRUE, the we will obscure these errors. But that all needs more
design anyway so I'm not going to stress about it yet.