spellcheck: start on spellcheck plugin
This starts bringing back the spellcheck plugin using our more
simplified APIs in libide. We still need to bring back the
widgets and transient sidebar features.

Additionally, we've removed the bubble drawing from source view
and we should probably add something that allows us to re-enable
that in a generic fashion.