debugger: add IdeDebugger API
This provides a somewhat simplistic API abstracting debugger communication.
A reference implementation will be provided using GDB.

There are some things still missing, but we can add that as we go. In
particular, we don't yet have a way to dive into complex variables and
tracking variable changes across runs (you have to requery, which is
potentially slow for various slow hardware).

We still need to get the UI in place, but this has some reusable components
that can be used in the debugger perspective.